lib,test,tools: remove unneeded block statements
Arrow functions that are a single expression do not need to be their own
blocks. Replace instances of single expressions as blocks in arrow
functions in the code base.

This is in preparation for enabling a lint rule to enforce this
practice.

The "use or do not use blocks" discussion came up in a pull request.
This change will codify the prevalent practice in the code base in a
lint rule.