doc: update deprecations info in COLLABORATOR_GUIDE
The deprecations section in the COLLABORATOR_GUIDE is a bit long-winded
and repetitive. It also includes some aspirational items that appear
worded to create "an impression that a specific or meaningful statement
has been made, when instead only a vague or ambiguous claim has actually
been communicated." (Putting it in quotes because I'm copying it from
Wikipedia.)

This commit edits the text to make it more concise and meaningful.